editor: Fix sometines too big initial window size.
Previously the initial window size was a fixed to whatever system of
measure JUCE uses (it is scaling dependant).

The previous code was wrongly assuming that those coordinates could fit
the plugin. It wasn't the case on some user's screens.

This commit sets the initial window size based on a ratio of the
available screen size.