mixmaxtrix: add module oversample class
This new "oversampled" class does the boilerplate of oversampling a
(probably 3rd party) module.

On this class the coeffients for acceptable realtime oversampling FIR
filters are introduced, aiming for 32 taps for each polyphase branch, so
a latency of 32 samples when enabled. The filter has the -6dB cutoff at
Nyquist and a 4KHz passband, so at the worst case of 44.1kHz aliasing increases from -96dB at 18KHz upwards to 22kHz. Hopefully this is just
acceptable, as aliasing falls in a range that most people can't hear
with very low energy in audio signals.

Using r8brainfree has to be investigated. Obvously it is done by a DSP
expert. I'd have to check:

- At which latencies and CPU costs operates.
- How much bloat I get by adding another FFT to the project.
- If adding the intel IPP FFT is Github CI friendly (I guess not).

The current FIR resamplers use float. Trying to see quantization errors
on such short filters become noticeable is to be done. A possible
shortcut, measuring CPU usage increases on float (memory requirement
increase would be hard to measure).

Lth band optimizations on the FIR filter are to be done.