mixmaxtrix: processor: refresh parameters only once
This commit moves the parameter refresh (polling from the Juce APVTS
atomic pointers to local copies) so it is all done on the same place
and in contiguous order.

This doesn't apply for fx parameters, as there are hundreds and only
a small subset is simultaneously required. Those are still done per
channel basis.
)