saturation: reduce CPU consumption
This commit improves CPU consumption.

SSE is added but only applied to sqrt, as it didn't show a lot of
benefit on sqrt2 and hardclip.

Most of the gains come from realizing that "constexpr_db_to_gain" was
called at data rate, while a constexpr function called with a constant
could have been removed by the compiler.