xsimd: wrap workarounds on macro
All xsimd workarounds because of ffast-math brokeness are wrapped on
the XSIMD_BROKEN_W_FAST_MATH macro. This may allow an easy revert.

Ideally every xsimd operation should be wrapped and there make it to
work with scalars, but that is a lot of work. This is good enough and
still clean.