xsimd: replace with compiler vector types
The XSIMD fixes for broken functions when -ffast-math is enabled came
with interface changes too. The xsimd::batches are now fixed to an
architecture class instead of just giving the size, so they lose a bit
of platform portability. Changes on all the usage sites would have been
required.

On this commit I went the extra mile and replaced XSIMD by the native
compiler vector types. XSIMD is still used to provide implementations
of math functions, but not as a primary wrapper type. The function
implementations are also wrapped.

The rationale for using the vector types are:

- The interface will be stable. No more breaking changes.

- As they are seen by the compiler, it allows it to hopefully make
  smarter decisions.

- Operations with scalar don't require constructing batches depending
  on the operation order.

Unfortunately:

- The vector types use __attribute__,  attributes can't be used as
  regular constructs of the language, so templates can't infer the
  "vector_size(x)" values. This has been workarounded by creating a
  traits class that extracts most data of a vector register.

- As I can't use templates on the "vector_size" __attribute__, vector
  types are passed as a single type on all templates, which results in
  less readable code and more SFINAE.

- Support on GCC is even worse as of now and vector types don't
  participate on overload resolution. This breaks the code for GCC until
  they fix it (8 years old bug).
  https://gcc.gnu.org/bugzilla//show_bug.cgi?id=57572

- < Clang 10 has the same problem than GCC, so Clang 11 is required.
  Luckily Visual Studio's 2019 bundles clang-cl 11.