simd: Introduce vectors of one element.
Hopefully these will be used for removing all the multiple "_simd" functions, which gets very annoying.

If a scalar version is desired the vector functions will be possible
to call by wrapping the parameters through "make_vec_x1".

Ideally, single coefficient versions should be dropped, but they save
runtime memory (cache), so for those special casing and implementation duplication could be kept, as the implementation would be always SIMD,
all the class has to do is to implement a single function that gets
the loaded coefficients as function parameters.