Corrupted git tree commit.
The computer crashed while building and the .git tree was corrupt beyond
repair. Fortunately the local copy was intact.

No operation was going on with Git, just building. Which is weird. Just
VSCode, Firefox and "make" were running.

So individual commits are lost. From top of my head since last update
on the remote this happened:

- FIR CPU improvements.
- filter2x CPU improvements.
- filter2x got an envelope follower.
- Compiler warnings removed.
- More small changes.

Not redoing the individual commits.