xsimd: Bump
Most of the functions are fixed now on a patch I submitted. They did
a patch but it didn't work on Clang.

Just atan2 is left.

This commit also addds tests again std for all defined functions.

BUT, by measuring it seems that XSIMD is actually slower than 2x std
on this use case. So it is kept disabled.