mixmaxtrix: Make crossover an FX
This commit hacks a special method for the crossover send outputs to
other channels.

The rationale for this is:

- It allows using only channel 1 and leaving 7 channels for FX free.
  This is a big win interface-wise.
- It opens the door for a 7 band crossover (limited by param count),
  should I want it.
- It doesn't require much special handling on the editor. Quite the
  contrary logic is removed.
- It is still implemented as a hack, as the previous variant. Nothing
  lost.

The approach taken is to add 3 extra slots to the mixing inputs for each
of the 8 channels, these ALWAYS belong to the crossover outputs 2 3 and
4. There is only one crossover instance available at channel 0.

Reordering optimizations are inhibited on these channels.