Add Dry Wet difference sends
This commit allows a channel to send the difference between its internal
Dry and Wet busses to the neighboring channel.

This was done after realizing that the transient splitter could be
converted into a transient gate that can operate on each channel by
using this feature. Much nicer than to have it as another crossover
hack.