parts_to_class: Add variant for different types.
This adds a variant of parts to class that allows containing
different types.

This will hopefully enable writing fx in a more clear way. The
codebase readability is paying a high price for having the states
separated from the functions. I still consider it desirable, but
will need some tooling.

Ideally a lot of the pointer offsetting should be done by the
compiler.