reversed cpole: Interleave delay lines
This commit uses the complex class for computations, making for much
tidier code.

A consequence of storing complex types now is that the real and
imaginary parts are interleaved on the buffer. It seems to give a
slight performane increase but it wasn't measured seriously because
I just was looking for performance regressions, as the goal of the
patch was to make the code more clear by using "vec_complex".