fft: Add pfft
WDL's FFT seems to only accept sizes up to 32K elements. The upcoming
resampler might require FFT's bigger than that. This commit introduces
PFFFT as a replacement of for FFT's of double precision. So mufft will
be used for "float" and "pffft" for double.

The inteface of the FFT's is also improved.