widgets: combobox: Fix constructed value
On construction of the combobox, Windows seems to be setting the value
on the constructor itself (via the attachment). On Linux it is done
after the constructor is concluded. So not setting the default value
left Linux in an undefined state. On Windows setting the value
unconditionally overwrote the value that it got via the attachment.