lofiverb: Go back to 16-bit fixed point
The sound differences between allowing the fixed point math go up to
64 bits and just truncating at 16-bit for storage were not big enough
to have separated.

This commit limits the intermediate math to 32 bit. The good thing is
that this is done while keeping the same code than float, enabled by the
extensive work done on the fixed point and ratio classes.